Version 1.20 alpha 2
Removed define of printf=iprintf in build settings
Removed Platform::Time() function, use millis() or millis64() instead
Added event logging to SD card. MessageType is now a bitmap. Implemented
M929.
Enable pullup resistors on endstops 10 and 11
SCARA printers can now use the manual bed levelling assistant
When using the manual bed levelling assitant, don't give an error if the
bed screw corrections were out of range, and always leave the first
screw alone
M552 now supports connection to a specified SSID
M572 now allows multiple D values
Thermocouple type letter in M305 command may now be in lower case
Bug fix: G29 bed probing on a SCARA printer gave spurious "not
reachable" warnings and skipped probe points
Protect against a dud command line containing the letter M being
interpreted as a M0 command
Fix reference to towers in the error message when trying to move a SCARA
printer before homing it
When updating firmware, warn that USB will disconnect
Fix duplicate error messatge when opening a gcode file fails

+++ src/Duet/Network.h
@@ -115,7 +115,7 @@ public:
 private:
 	Platform& platform;
 	Webserver *webserver;
-	float longWait;
+	uint32_t longWait;
 
 	void AppendTransaction(NetworkTransaction* volatile * list, NetworkTransaction *r);
 	void PrependTransaction(NetworkTransaction* volatile * list, NetworkTransaction *r);
